requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
docker==5.0.0
flake8==3.9.2
flake8-quotes==3.2.0
gitchangelog==3.0.4
gitdb==4.0.7
GitPython==3.1.17
glob2==0.7
idna==2.10
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
parse==1.19.0
parse-type==0.5.2
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pystache==0.5.4
pytest==6.2.4
pytest-bdd==4.0.2
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
smmap==4.0.0
toml==0.10.2
urllib3==1.26.5
websocket-client==1.1.0
yamllint==1.26.1
